convert all sensor putputs to double precision; let all sensors have their own scale factor gene that scales the global intScaleFactor to amplify outputs

+++ src/de/hansinator/fun/jgp/life/SensorInput.java
@@ -9,5 +9,5 @@ public interface SensorInput {
 
 	public static final SensorInput[] emptySensorInputArray = {};
 
-	public int get();
+	public double get();
 }
